Add heap task implementation
New file heap_task.py is created, where a list of random numbers is generated and manipulated using the heapq module. The main operation carried out in this code is the creation of a heap from a list, followed by repeated removal and addition of elements to find the minimum possible total cost.